requirements.txt
aniso8601==7.0.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.1.2
Flask-GraphQL==2.0.1
Flask-HTTPAuth==4.7.0
Flask-SeaSurf==1.1.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
httplib2==0.20.4
idna==3.4
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
promise==2.3
psycopg==3.0.14
pyparsing==3.0.9
python-dotenv==1.0.0
requests==2.31.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.36
urllib3==2.0.3
Werkzeug==2.1.2
zipp==3.8.0
